Move package logic into __init__.py
+ Leaves arxiv.py as a deprecated alias.
+ Updates docs target accordingly; enable search.
