+++ Makefile
@@ -20,9 +20,7 @@ audit:
 docs: docs/index.html
 docs/index.html: $(source) README.md
 	pdoc --version
-	pdoc --docformat "restructuredtext" ./arxiv/arxiv.py -o docs --no-search
-	mv docs/arxiv/arxiv.html docs/index.html
-	rmdir docs/arxiv
+	pdoc --docformat "restructuredtext" ./arxiv/__init__.py -o docs
 
 clean:
 	rm -rf build dist
